fastflow_training/test_requirements.txt
timm==0.5.4
FrEIA
torch==1.9.1
torchvision==0.10.1
pyyaml==6.0
pytorch-ignite==0.2.0
scikit-learn==0.24.2
